use decimal for number

+++ models/get_last_trades_by_instrument_and_time_params.go
@@ -2,8 +2,8 @@ package models
 
 type GetLastTradesByInstrumentAndTimeParams struct {
 	InstrumentName string `json:"instrument_name"`
-	StartTimestamp int    `json:"start_timestamp"`
-	EndTimestamp   int    `json:"end_timestamp"`
+	StartTimestamp uint64 `json:"start_timestamp"`
+	EndTimestamp   uint64 `json:"end_timestamp"`
 	Count          int    `json:"count,omitempty"`
 	IncludeOld     bool   `json:"include_old,omitempty"`
 	Sorting        string `json:"sorting,omitempty"`
